@@ -14,32 +14,31 @@
   "devDependencies": {
     "@floating-ui/dom": "^1.0.1",
     "@popperjs/core": "^2.11.6",
-    "@sveltejs/adapter-auto": "^1.0.0-next.86",
-    "@sveltejs/adapter-node": "^1.0.0-next.100",
+    "@sveltejs/adapter-auto": "next",
+    "@sveltejs/adapter-node": "next",
     "@sveltejs/kit": "^1.0.0-next.532",
     "@types/lz-string": "^1.3.34",
-    "@typescript-eslint/eslint-plugin": "^5.27.0",
-    "@typescript-eslint/parser": "^5.27.0",
+    "@typescript-eslint/eslint-plugin": "^5.45.0",
+    "@typescript-eslint/parser": "^5.45.0",
     "autoprefixer": "^10.4.8",
     "classnames": "^2.3.2",
-    "eslint": "^8.16.0",
-    "eslint-config-prettier": "^8.3.0",
+    "eslint": "^8.28.0",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-svelte3": "^4.0.0",
     "firebase-admin": "^11.2.0",
     "flowbite": "^1.5.3",
     "flowbite-svelte": "^0.27.11",
     "postcss": "^8.4.16",
-    "prettier": "^2.6.2",
-    "prettier-plugin-svelte": "^2.7.0",
-    "svelte": "^3.44.0",
-    "svelte-check": "^2.7.1",
-    "svelte-heros": "^2.3.5",
+    "prettier": "^2.8.0",
+    "prettier-plugin-svelte": "^2.8.1",
+    "svelte": "^3.53.1",
+    "svelte-check": "^2.9.2",
     "svelte-heros-v2": "^0.2.2",
-    "svelte-preprocess": "^4.10.6",
+    "svelte-preprocess": "^4.10.7",
     "tailwindcss": "^3.1.8",
-    "tslib": "^2.3.1",
-    "typescript": "^4.7.4",
-    "vite": "^3.1.0"
+    "tslib": "^2.4.1",
+    "typescript": "^4.9.3",
+    "vite": "^3.2.4"
   },
   "type": "module",
   "dependencies": {
